Move into utility file
